Implemented sliding window attention to maintain KV cache only for the window size to enable infinite decoding.
Currently, when using `MultiheadAttention` or `GroupedQueryAttention` for
sliding window attention, the KV cache is kept for the full sequence length
(`seq_len`) instead of the window length (`window_len`).

For example, a model with `window_len=1k` and `seq_len=2M` keeps a KV cache
for the full 2M tokens. It then biases 1999k invalid KV tokens before
calculating attention, resulting in a computational complexity of **O(2M²)**
instead of the desired **O(1k²)**.

This issue persists even when using flash attention. Flash attention uses the
KV cache allocated in HBM as its input. While unnecessary blocks are discarded
during computation, the KV cache still occupies HBM inefficiently for the full
2M tokens.

To address this, when `MultiheadAttention` detects a sliding window mask, it
stores the key-value (KV) cache in a ring buffer inside the input linear layer.
As a result, downstream projects using `MultiheadAttention` automatically
benefit from efficient KV cache handling in `init_states` and `extend_step`.

Additionally, for use cases like local-global attention in LLMs, it is
recommended to use sliding window masks for even the global attention as well.
For example, if you want to train an LLM with a context length of 8k, you can
set the sliding window size to 8k during training. This enables functionally
infinite decoding during inference. Accuracy wouldn't be good tho.

Note:
* query_positions in QKVLinear.forward() was introduced by
  https://github.com/apple/axlearn/pull/914. Now it returns to the caller.

This PR actually moves from downstream speech/streaming/sliding_window_attention.py

+++ axlearn/common/encoder.py
@@ -174,7 +174,7 @@ class CausalEncoder(Encoder):
         cfg: CausalEncoder.Config = self.config
         init_state, _ = self.transformer.init_states(
             time_step=None,
-            data=TensorSpec([batch_size, max_sequence_length, cfg.dim]),
+            data=TensorSpec([batch_size, max_sequence_length, cfg.dim], dtype=jnp.float32),
         )
         return dict(
             transformer_state=init_state,
